refactor(feature:path-tracking): migrate to CMP
