+++ feature/path-tracking/src/nativeMain/kotlin/com/mifos/feature/path/tracking/PathTrackingScreen.native.kt
@@ -10,6 +10,7 @@
 package com.mifos.feature.path.tracking
 
 import androidx.compose.runtime.Composable
+import com.mifos.core.model.objects.users.UserLatLng
 import com.mifos.feature.pathTracking.PathTrackingViewModel
 
 @Composable
@@ -17,5 +18,20 @@ actual fun PathTrackingScreen(
     onBackPressed: () -> Unit,
     viewModel: PathTrackingViewModel,
 ) {
+    TODO("Not yet implemented")
+}
+
+@Composable
+actual fun HandleLocationPermissionRequest(
+    show: Boolean,
+    onPermissionResult: (granted: Boolean) -> Unit,
+) {
+    if (show) {
+        onPermissionResult(true)
+    }
+}
+
+@Composable
+actual fun PathTrackingMapView(latLngList: List<UserLatLng>) {
     TODO("Not yet implemented")
 }
